JOY-86: rename folder
