+++ src/features/profile/ui/profileSettings/ProfileSettings.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 
 import { Tabs } from '@/shared/ui'
 import { Tab } from '@/shared/ui/tabs'
-import { GeneralInformation } from '../generalInformationTab/GeneralInformation'
+import { GeneralInformation } from '../generalInformation/GeneralInformation'
 import { PATH } from '@/shared/config/routes'
 
 import s from './ProfileSettings.module.scss'
